move basicAuth, tokenAuth and checkConnection to Adaptor.js
