+++ packages/salesforce/src/util.js
@@ -1,51 +1,4 @@
-import { Connection } from '@jsforce/jsforce-node';
 import { throwError } from '@openfn/language-common/util';
-
-const checkConnection = connection => {
-  if (!connection) {
-    throwError('No connection');
-  }
-  console.log('Using Salesforce API version:', connection.version);
-  console.log(`Connected with ${connection._sessionType} session type`);
-};
-
-export const basicAuth = async configuration => {
-  const {
-    loginUrl,
-    username,
-    password,
-    securityToken,
-    apiVersion: version,
-  } = configuration;
-
-  const conn = new Connection({ loginUrl, version });
-  console.info(`Connecting to salesforce as ${username}.`);
-
-  await conn
-    .login(username, securityToken ? password + securityToken : password)
-    .catch(error => {
-      throwError('FAILED_AUTH', {
-        fix: 'Check your username, password, and security token',
-        message: `Failed to connect to salesforce as ${username}`,
-        error,
-      });
-    });
-  checkConnection(conn);
-
-  return conn;
-};
-
-export const tokenAuth = configuration => {
-  const {
-    instance_url: instanceUrl,
-    access_token: accessToken,
-    apiVersion: version,
-  } = configuration;
-
-  const conn = new Connection({ instanceUrl, accessToken, version });
-  checkConnection(conn);
-  return conn;
-};
 
 let anyAscii = undefined;
